feat: addition of account collector tests and target membership
